fixed up 68578072bfaf6054a96bb6bcedfccb6e56a508fe: negative numbers weren't parsed correctly

+++ src/expressions.c
@@ -119,7 +119,7 @@ teco_expressions_add_digit(gunichar digit)
 	teco_int_t n = teco_expressions_args() > 0 ? teco_expressions_pop_num(0) : 0;
 
 	/* use g_unichar_digit_value()? */
-	teco_expressions_push(n*teco_radix + (n < 0 ? -1 : 1)*(digit - '0'));
+	teco_expressions_push(n*teco_radix + (n < 0 ? -1 : 1)*((gint)digit - '0'));
 }
 
 void
